GType is a gulong, not a guint. Fixes crash on x86_64. #500508
	* codegen/defsgen.py (GType): GType is a gulong, not a guint.
	Fixes crash on x86_64. #500508

svn path=/trunk/; revision=2932

+++ codegen/defsgen.py
@@ -489,7 +489,7 @@ class DefsWriter:
 
 # ---------- ctypes support classes for gobject library functions ----------
 
-GType = ctypes.c_uint
+GType = ctypes.c_ulong
 
 class GTypeClass(ctypes.Structure):
     _fields_ = [('g_type', GType)]
